Added military-hardware
This is the very first mode I wrote.  It does not yet meet the criteria of being "well written" and "using modern practices", but at least it is finally in source control.

Side note:  This is the first mode ever written pulling options from a config file.  Hence it does not have an ini-system included per-se, but does manually read an ini-like file for options.